Rename Test files
* Rename test classes to end with Test for exclusion from analysis
* Rename test suites to end in TestSuite
