+++ domain/src/test/java/org/sagebionetworks/research/domain/impl/ActionGsonTestSuite.java
@@ -32,35 +32,11 @@
 
 package org.sagebionetworks.research.domain.impl;
 
-import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
 
-public class ActionImplTests extends IndividualActionTests {
-    public static final ActionImpl COMPLETE = ActionImpl.builder().setButtonIconName("icon").setButtonTitle("title")
-            .build();
-
-    public static final ActionImpl NO_TITLE = ActionImpl.builder().setButtonIconName("icon").build();
-
-    public static final ActionImpl NO_ICON = ActionImpl.builder().setButtonTitle("title").build();
-
-    public static final ActionImpl EMPTY = ActionImpl.builder().build();
-
-    @Test
-    public void testAction_Complete() {
-        this.testCommon(COMPLETE, "Action_Complete.json");
-    }
-
-    @Test
-    public void testAction_Empty() {
-        this.testCommon(EMPTY, "Action_Empty.json");
-    }
-
-    @Test
-    public void testAction_NoIcon() {
-        this.testCommon(NO_ICON, "Action_NoIcon.json");
-    }
-
-    @Test
-    public void testAction_NoTitle() {
-        this.testCommon(NO_TITLE, "Action_NoTitle.json");
-    }
+@RunWith(Suite.class)
+@SuiteClasses({ActionImplTest.class, ReminderActionImplTest.class, SkipToStepActionImplTest.class})
+public class ActionGsonTestSuite {
 }
