+++ domain/src/test/java/org/sagebionetworks/research/domain/navigation/IndividualNavigatorTest.java
@@ -23,7 +23,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-public abstract class IndividualNavigatorTests {
+public abstract class IndividualNavigatorTest {
     private static final String STEP_TYPE = "step";
 
     protected static final List<Step> TEST_STEPS;
@@ -94,7 +94,7 @@ public abstract class IndividualNavigatorTests {
         return taskResult;
     }
 
-    public IndividualNavigatorTests(StepNavigator navigator) {
+    public IndividualNavigatorTest(StepNavigator navigator) {
         this.navigator = navigator;
         this.steps = TEST_STEPS;
     }
