+++ domain/src/test/java/org/sagebionetworks/research/domain/navigation/NavigatorTestSuite.java
@@ -5,7 +5,7 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses({TreeNavigatorTests.class, StrategyBasedNavigatorTests.class})
-public class AllNavigatorTests {
+@SuiteClasses({TreeNavigatorTest.class, StrategyBasedNavigatorTest.class})
+public class NavigatorTestSuite {
 
 }
